Track libraries as submodules.
There's quite a few dependencies, mostly because of go-cid -- which
the go-ipld-prime dagjson package depends on.

We probably actually want to make the JST package *not* depend on
the go-ipld-prime dagjson package specifically to avoid this weight.
That will require some refactoring, though, and may result in some
near-duplicate code which varies only in whether or not it handles
CIDs. (Such near-duplicate code will be unfortunate, but given how
ridiculously high-cost 'go-cid' is as a dependency, it may be a valid
price to pay for getting reasonably minimized dependency sprawl.)
This splitting is for a later date, though.